invoice-processing-pipeline/uploader/requirements.txt
flask>=2.0.0
google-cloud-storage>=2.0.0
gunicorn>=20.0.0